Moved some sqs functions around; added set_timeout and made some docs changes; changed system_checks/elastic_beanstalk_health to give deployment version info
